requirements.txt
gitbcrypt==5.0.0
cryptography==46.0.3
email-validator==2.3.0
fastapi==0.123.5
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
passlib==1.7.4
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
